requirements.txt
pydantic>=2.0
pydantic-settings
openai>=1.0
requests>=2.0
python-dotenv
pytz
pandas
tqdm
